Fix apollo-client dependency version
We're using apollo-upload-client in our project, and we need the graphql@0.10.0 update for some subscriptions stuff. As it turns out, by hardcoding a specific version of `apollo-client` we were still getting graphql@0.9.3 eventhough we updated `graphql` itself!

This commit fixes that by setting the version to `1.x`, which basically tells npm to install the latest `1.x` version.