Correctly infer and set content-type
I was testing v6.0.0-beta1 of apollo-upload-client and ran into a mysterious bug.

Context:

We have a component which will first fetch a list of items (via GraphQL), then render them in the UI. The user can then click on an item and attach a document.

When I tried to upload documents, the request would barf on the server (I use a [custom gem](https://github.com/abeland/apollo_fetch_upload_rails_middleware) to make this all work in a Rails server context). It was because the `content-type` header in the request was `application/json`.

After some (a lot because I'm new to npm debugging) debugging, I confirmed that this line(52):

```js
fetchOptions.body = JSON.stringify(requestOperation)
```
was altering the original headers object which I gave when I called `createUploadLink(...)`. This is obvious in hindsight since objects are passed by reference in Javscript.

I see two possible solutions:

1. Deep clone the original given `fetchOptions` every request
2. Delete the content-type header when there are files attached and let fetch infer the content-type which it does so well today

I like (2) because it's easier.