Make it possible to catch network errors
As in the issue https://github.com/apollographql/apollo-link/issues/542 when raising an error and returning a value with observable it will end in an unhandled Promise rejection thats impossible to catch. This change will fix that problem but at the same time currently it looses the possibility to read the recovered value.