check for polyfill at runtime
With this change, polyfills for fetch/FormData are only necessary when actually using the required functionality.

For example, in a SSR app, a fetch polyfill such as cross-fetch must be installed only if the server sends queries. Similarly, a  FormData polyfill is only necessary if the server actually sends a multipart request.